wip - more refactoring
