+++ rust/operator-binary/src/config/commands.rs
