+++ rust/operator-binary/src/config/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod product_logging;
+pub mod superset;
